code cleanup: structure and deduplication
