+++ NEShim/NEShim/Config/AppConfig.cs
@@ -120,16 +120,4 @@ public sealed class AppConfig
     public bool OverrideStartBindingProtection { get; set; } = false;
 }
 
-public sealed class InputBinding
-{
-    public string? Key           { get; set; }
-    public string? GamepadButton { get; set; }
-
-    public InputBinding() { }
-    public InputBinding(string? key, string? gamepadButton)
-    {
-        Key           = key;
-        GamepadButton = gamepadButton;
-    }
-}
