+++ NEShim/NEShim/Saves/SaveStateManager.cs
@@ -132,7 +132,3 @@ internal sealed class SaveStateManager
 
 }
 
-public sealed class SlotMeta
-{
-    public DateTime Timestamp { get; set; }
-}
