+++ NEShim/NEShim/Saves/SlotMeta.cs
@@ -1,0 +1,6 @@
+namespace NEShim.Saves;
+
+public sealed class SlotMeta
+{
+    public DateTime Timestamp { get; set; }
+}
