+++ NEShim/NEShim/UI/InGameMenu.cs
@@ -13,14 +13,8 @@ namespace NEShim.UI;
 /// Per-screen title, items, enabled state, and activation logic live in nested
 /// ScreenHandler classes — one per Screen enum value.
 /// </summary>
-internal sealed class InGameMenu
+internal sealed partial class InGameMenu
 {
-    public enum Screen
-    {
-        Root, SaveSlotSelect, Settings, KeyboardBindings, GamepadBindings,
-        Video, Sound, ConfirmLoad, ConfirmMainMenu, ConfirmExit, ControllerDisconnected
-    }
-
     private readonly SaveStateManager _saveStates;
     private readonly AppConfig        _config;
     private readonly LocalizationData _localization;
@@ -133,9 +127,15 @@ internal sealed class InGameMenu
             [Screen.GamepadBindings]        = new GamepadBindingsHandler(this),
             [Screen.Video]                  = new VideoHandler(this),
             [Screen.Sound]                  = new SoundHandler(this),
-            [Screen.ConfirmLoad]            = new ConfirmLoadHandler(this),
-            [Screen.ConfirmMainMenu]        = new ConfirmMainMenuHandler(this),
-            [Screen.ConfirmExit]            = new ConfirmExitHandler(this),
+            [Screen.ConfirmLoad]            = new ConfirmHandler(this,
+                _localization.InGameLoadTitle,   _localization.InGameConfirmYesLoad,
+                () => { _saveStates.LoadFromActiveSlot(); Close(); }),
+            [Screen.ConfirmMainMenu]        = new ConfirmHandler(this,
+                _localization.InGameReturnTitle, _localization.InGameConfirmYesReturn,
+                () => { Close(); _onReturnToMainMenu(); }),
+            [Screen.ConfirmExit]            = new ConfirmHandler(this,
+                _localization.InGameExitTitle,   _localization.InGameConfirmYesExit,
+                () => { Close(); _onExitToDesktop(); }),
             [Screen.ControllerDisconnected] = new ControllerDisconnectedHandler(this),
         };
 
@@ -390,282 +390,4 @@ internal sealed class InGameMenu
 
     public void Render(Graphics g, Rectangle bounds) => MenuRenderer.Draw(g, bounds, this);
 
-    // =========================================================================
-    // Screen handlers
-    // Each handler encapsulates one screen's title, items, enabled state, and
-    // activation logic. To add a new screen: add an enum value, a handler class,
-    // and an entry in BuildHandlers().
-    // =========================================================================
-
-    private abstract class ScreenHandler
-    {
-        protected InGameMenu Menu { get; }
-        protected ScreenHandler(InGameMenu menu) => Menu = menu;
-        public abstract string   Title     { get; }
-        public abstract int      ItemCount { get; }
-        public abstract string[] GetItems();
-        public abstract void     Activate(int index);
-        public virtual  bool     IsItemEnabled(int index) => true;
-    }
-
-    private sealed class RootHandler : ScreenHandler
-    {
-        public RootHandler(InGameMenu menu) : base(menu) { }
-        public override string Title     => Menu._localization.InGamePausedTitle;
-        public override int    ItemCount => 8;
-        public override string[] GetItems() => new[]
-        {
-            Menu._localization.InGameResume,
-            Menu._localization.InGameResetGame,
-            Menu._localization.InGameSelectSaveSlot,
-            Menu._localization.InGameSaveGame,
-            Menu._localization.InGameLoadGame,
-            Menu._localization.InGameSettings,
-            Menu._localization.InGameReturnToMain,
-            Menu._localization.InGameExit,
-        };
-        public override bool IsItemEnabled(int index) =>
-            index != RootItemLoadGame || Menu._saveStates.SlotExists(Menu._saveStates.ActiveSlot);
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case 0:                    Menu.Close(); break;
-                case 1:                    Menu._onResetGame(); Menu.Close(); break;
-                case 2:                    Menu.NavigateTo(Screen.SaveSlotSelect); break;
-                case 3:                    Menu._saveStates.SaveToActiveSlot(); Menu.Close(); break;
-                case RootItemLoadGame:     Menu.NavigateTo(Screen.ConfirmLoad); break;
-                case 5:                    Menu.NavigateTo(Screen.Settings); break;
-                case RootItemReturnToMain:
-                    Menu.NavigateTo(Screen.ConfirmMainMenu);
-                    Menu.SelectedItem = 1;
-                    break;
-                case 7:
-                    Menu.NavigateTo(Screen.ConfirmExit);
-                    Menu.SelectedItem = 1;
-                    break;
-            }
-        }
-    }
-
-    private sealed class SaveSlotSelectHandler : ScreenHandler
-    {
-        public SaveSlotSelectHandler(InGameMenu menu) : base(menu) { }
-        public override string Title =>
-            string.Format(Menu._localization.InGameSelectSlotTitle, Menu._saveStates.ActiveSlot + 1);
-        public override int ItemCount => SaveStateManager.SlotCount + 1;
-        public override string[] GetItems()
-            => Enumerable.Range(0, SaveStateManager.SlotCount)
-                .Select(i => string.Format(Menu._localization.SlotLabel, i + 1)
-                           + (i == Menu._saveStates.ActiveSlot ? Menu._localization.SlotActive : ""))
-                .Append(Menu._localization.Back)
-                .ToArray();
-        public override void Activate(int index)
-        {
-            if (index == SaveStateManager.SlotCount)
-                Menu.NavigateTo(Screen.Root);
-            else
-            {
-                Menu._saveStates.ActiveSlot = index;
-                Menu._config.ActiveSlot     = index;
-                Menu.NavigateTo(Screen.Root);
-            }
-        }
-    }
-
-    private sealed class SettingsHandler : ScreenHandler
-    {
-        public SettingsHandler(InGameMenu menu) : base(menu) { }
-        public override string   Title     => Menu._localization.SettingsTitle;
-        public override int      ItemCount => 5;
-        public override string[] GetItems() => new[]
-        {
-            Menu._localization.SettingsKeyboard,
-            Menu._localization.SettingsGamepad,
-            Menu._localization.SettingsVideo,
-            Menu._localization.SettingsSound,
-            Menu._localization.Back,
-        };
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case 0: Menu.NavigateTo(Screen.KeyboardBindings); break;
-                case 1: Menu.NavigateTo(Screen.GamepadBindings);  break;
-                case 2: Menu.NavigateTo(Screen.Video);            break;
-                case 3: Menu.NavigateTo(Screen.Sound);            break;
-                case 4: Menu.NavigateTo(Screen.Root);             break;
-            }
-        }
-    }
-
-    private sealed class KeyboardBindingsHandler : ScreenHandler
-    {
-        public KeyboardBindingsHandler(InGameMenu menu) : base(menu) { }
-        public override string Title => Menu.RebindingAction != null
-            ? string.Format(Menu._localization.PressKeyTitle,
-                Menu._bindingActions.First(b => b.ConfigKey == Menu.RebindingAction).Label.ToUpper())
-            : Menu._localization.SettingsKeyboard.ToUpper();
-        public override int      ItemCount => Menu._bindingActions.Length;
-        public override string[] GetItems()
-            => Menu._bindingActions
-                .Select(b => b.ConfigKey == ""
-                    ? Menu._localization.Back
-                    : $"{b.Label,-8}  {Menu.KeyboardLabel(b.ConfigKey)}")
-                .ToArray();
-        public override void Activate(int index)
-        {
-            var (_, configKey) = Menu._bindingActions[index];
-            if (configKey == "")
-                Menu.NavigateTo(Screen.Settings);
-            else
-                Menu.RebindingAction = configKey;
-        }
-    }
-
-    private sealed class GamepadBindingsHandler : ScreenHandler
-    {
-        public GamepadBindingsHandler(InGameMenu menu) : base(menu) { }
-        public override string Title => Menu.GamepadRebindingAction != null
-            ? string.Format(Menu._localization.PressButtonTitle,
-                Menu._gamepadBindingActions.First(b => b.ConfigKey == Menu.GamepadRebindingAction).Label.ToUpper())
-            : Menu._localization.SettingsGamepad.ToUpper();
-        public override int      ItemCount => Menu._gamepadBindingActions.Length;
-        public override string[] GetItems()
-            => Menu._gamepadBindingActions
-                .Select(b => b.ConfigKey == ""
-                    ? Menu._localization.Back
-                    : $"{b.Label,-8}  {Menu.GetGamepadLabel(b.ConfigKey)}")
-                .ToArray();
-        public override bool IsItemEnabled(int index)
-        {
-            if (!SteamInputManager.IsUsingNativeActions()) return true;
-            var configKey = Menu._gamepadBindingActions[index].ConfigKey;
-            return configKey == "" || configKey == "OpenMenu";
-        }
-        public override void Activate(int index)
-        {
-            var (_, configKey) = Menu._gamepadBindingActions[index];
-            if (configKey == "")
-                Menu.NavigateTo(Screen.Settings);
-            else
-                Menu.GamepadRebindingAction = configKey;
-        }
-    }
-
-    private sealed class VideoHandler : ScreenHandler
-    {
-        public VideoHandler(InGameMenu menu) : base(menu) { }
-        public override string   Title     => Menu._localization.VideoTitle;
-        public override int      ItemCount => 4;
-        public override string[] GetItems() => new[]
-        {
-            Menu._config.WindowMode == "Fullscreen" ? Menu._localization.VideoWindowFullscreen : Menu._localization.VideoWindowWindowed,
-            Menu._config.GraphicsSmoothingEnabled   ? Menu._localization.VideoGraphicsSmooth   : Menu._localization.VideoGraphicsOriginal,
-            Menu._config.ShowFps                    ? Menu._localization.VideoFpsOn            : Menu._localization.VideoFpsOff,
-            Menu._localization.Back,
-        };
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case 0:
-                    Menu._onWindowModeToggle(Menu._config.WindowMode != "Fullscreen");
-                    break;
-                case 1:
-                    bool smoothOn = !Menu._config.GraphicsSmoothingEnabled;
-                    Menu._config.GraphicsSmoothingEnabled = smoothOn;
-                    Menu._onGraphicsScalerToggled(smoothOn);
-                    break;
-                case 2:
-                    Menu._config.ShowFps = !Menu._config.ShowFps;
-                    Menu._onConfigSaved();
-                    break;
-                case 3:
-                    Menu.NavigateTo(Screen.Settings);
-                    break;
-            }
-        }
-    }
-
-    private sealed class SoundHandler : ScreenHandler
-    {
-        public  const int VolumeIndex   = 0;
-        private const int ScrubberIndex = 1;
-        private const int BackIndex     = 2;
-        public SoundHandler(InGameMenu menu) : base(menu) { }
-        public override string   Title     => Menu._localization.SoundTitle;
-        public override int      ItemCount => 3;
-        public override string[] GetItems() => new[]
-        {
-            string.Format(Menu._localization.SoundVolume, Menu._config.Volume),
-            Menu._config.SoundScrubberEnabled ? Menu._localization.SoundScrubberOn : Menu._localization.SoundScrubberOff,
-            Menu._localization.Back,
-        };
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case ScrubberIndex:
-                    bool scrubOn = !Menu._config.SoundScrubberEnabled;
-                    Menu._config.SoundScrubberEnabled = scrubOn;
-                    Menu._onScrubberToggled(scrubOn);
-                    break;
-                case BackIndex:
-                    Menu.NavigateTo(Screen.Settings);
-                    break;
-            }
-        }
-    }
-
-    private sealed class ConfirmLoadHandler : ScreenHandler
-    {
-        public ConfirmLoadHandler(InGameMenu menu) : base(menu) { }
-        public override string   Title     => Menu._localization.InGameLoadTitle;
-        public override int      ItemCount => 2;
-        public override string[] GetItems() =>
-            new[] { Menu._localization.InGameConfirmYesLoad, Menu._localization.InGameConfirmNoStay };
-        public override void Activate(int index)
-        {
-            if (index == 0) { Menu._saveStates.LoadFromActiveSlot(); Menu.Close(); }
-            else Menu.NavigateTo(Screen.Root);
-        }
-    }
-
-    private sealed class ConfirmMainMenuHandler : ScreenHandler
-    {
-        public ConfirmMainMenuHandler(InGameMenu menu) : base(menu) { }
-        public override string   Title     => Menu._localization.InGameReturnTitle;
-        public override int      ItemCount => 2;
-        public override string[] GetItems() =>
-            new[] { Menu._localization.InGameConfirmYesReturn, Menu._localization.InGameConfirmNoStay };
-        public override void Activate(int index)
-        {
-            if (index == 0) { Menu.Close(); Menu._onReturnToMainMenu(); }
-            else Menu.NavigateTo(Screen.Root);
-        }
-    }
-
-    private sealed class ConfirmExitHandler : ScreenHandler
-    {
-        public ConfirmExitHandler(InGameMenu menu) : base(menu) { }
-        public override string   Title     => Menu._localization.InGameExitTitle;
-        public override int      ItemCount => 2;
-        public override string[] GetItems() =>
-            new[] { Menu._localization.InGameConfirmYesExit, Menu._localization.InGameConfirmNoStay };
-        public override void Activate(int index)
-        {
-            if (index == 0) { Menu.Close(); Menu._onExitToDesktop(); }
-            else Menu.NavigateTo(Screen.Root);
-        }
-    }
-
-    private sealed class ControllerDisconnectedHandler : ScreenHandler
-    {
-        public ControllerDisconnectedHandler(InGameMenu menu) : base(menu) { }
-        public override string   Title     => "";
-        public override int      ItemCount => 0;
-        public override string[] GetItems() => Array.Empty<string>();
-        public override void     Activate(int index) { }
-    }
 }
