+++ NEShim/NEShim/UI/InGameMenuHandlers/Screen.cs
@@ -1,0 +1,10 @@
+namespace NEShim.UI;
+
+internal sealed partial class InGameMenu
+{
+    public enum Screen
+    {
+        Root, SaveSlotSelect, Settings, KeyboardBindings, GamepadBindings,
+        Video, Sound, ConfirmLoad, ConfirmMainMenu, ConfirmExit, ControllerDisconnected
+    }
+}
