+++ NEShim/NEShim/UI/MainMenuHandlers/ResumeOption.cs
@@ -1,0 +1,6 @@
+namespace NEShim.UI;
+
+internal sealed partial class MainMenuScreen
+{
+    private record ResumeOption(string Label, Action? Load);
+}
