+++ NEShim/NEShim/UI/MainMenuHandlers/Screen.cs
@@ -1,0 +1,6 @@
+namespace NEShim.UI;
+
+internal sealed partial class MainMenuScreen
+{
+    public enum Screen { Main, ResumeSlots, Settings, KeyboardBindings, GamepadBindings, Video, Sound }
+}
