+++ NEShim/NEShim/UI/MainMenuScreen.cs
@@ -14,10 +14,8 @@ namespace NEShim.UI;
 /// Per-screen title, items, enabled state, and activation logic live in nested
 /// ScreenHandler classes — one per Screen enum value.
 /// </summary>
-internal sealed class MainMenuScreen : IDisposable
+internal sealed partial class MainMenuScreen : IDisposable
 {
-    public enum Screen { Main, ResumeSlots, Settings, KeyboardBindings, GamepadBindings, Video, Sound }
-
     private readonly (string Label, string ConfigKey)[] _bindingActions;
     private readonly (string Label, string ConfigKey)[] _gamepadBindingActions;
     private readonly IReadOnlyDictionary<Screen, ScreenHandler> _handlers;
@@ -419,235 +417,5 @@ internal sealed class MainMenuScreen : IDisposable
         if (File.Exists(inCwd)) return inCwd;
 
         return null;
-    }
-
-    private record ResumeOption(string Label, Action? Load);
-
-    // =========================================================================
-    // Screen handlers
-    // Each handler encapsulates one screen's title, items, enabled state, and
-    // activation logic. To add a new screen: add an enum value, a handler class,
-    // and an entry in BuildHandlers().
-    // =========================================================================
-
-    private abstract class ScreenHandler
-    {
-        protected MainMenuScreen Menu { get; }
-        protected ScreenHandler(MainMenuScreen menu) => Menu = menu;
-        public abstract string   Title     { get; }
-        public abstract int      ItemCount { get; }
-        public abstract string[] GetItems();
-        public abstract void     Activate(int index);
-        public virtual  bool     IsItemEnabled(int index) => true;
-    }
-
-    private sealed class MainHandler : ScreenHandler
-    {
-        private const int ResumeIndex = 1;
-        public MainHandler(MainMenuScreen menu) : base(menu) { }
-        public override string   Title     => Menu._localization.MainMenuTitle;
-        public override int      ItemCount => 4;
-        public override string[] GetItems() => new[]
-        {
-            Menu._localization.MainMenuNewGame,
-            Menu._localization.MainMenuResumeGame,
-            Menu._localization.MainMenuSettings,
-            Menu._localization.MainMenuExit,
-        };
-        public override bool IsItemEnabled(int index) =>
-            index != ResumeIndex || Menu.CanResume;
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case 0:
-                    Menu.IsVisible = false;
-                    Menu.NewGameChosen?.Invoke();
-                    break;
-                case ResumeIndex:
-                    Menu.BuildResumeOptions();
-                    Menu.NavigateTo(Screen.ResumeSlots);
-                    break;
-                case 2:
-                    Menu.NavigateTo(Screen.Settings);
-                    break;
-                case 3:
-                    Menu.IsVisible = false;
-                    Menu.ExitChosen?.Invoke();
-                    break;
-            }
-        }
-    }
-
-    private sealed class ResumeSlotsHandler : ScreenHandler
-    {
-        public ResumeSlotsHandler(MainMenuScreen menu) : base(menu) { }
-        public override string   Title     => Menu._localization.MainMenuLoadTitle;
-        public override int      ItemCount => Menu._resumeOptions.Length;
-        public override string[] GetItems() => Menu._resumeOptions.Select(o => o.Label).ToArray();
-        public override void Activate(int index)
-        {
-            var opt = Menu._resumeOptions[index];
-            if (opt.Load == null)
-                Menu.NavigateTo(Screen.Main);
-            else
-            {
-                opt.Load();
-                Menu.IsVisible = false;
-                Menu.ResumeChosen?.Invoke();
-            }
-        }
-    }
-
-    private sealed class SettingsHandler : ScreenHandler
-    {
-        public SettingsHandler(MainMenuScreen menu) : base(menu) { }
-        public override string   Title     => Menu._localization.SettingsTitle;
-        public override int      ItemCount => 5;
-        public override string[] GetItems() => new[]
-        {
-            Menu._localization.SettingsKeyboard,
-            Menu._localization.SettingsGamepad,
-            Menu._localization.SettingsVideo,
-            Menu._localization.SettingsSound,
-            Menu._localization.Back,
-        };
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case 0: Menu.NavigateTo(Screen.KeyboardBindings); break;
-                case 1: Menu.NavigateTo(Screen.GamepadBindings);  break;
-                case 2: Menu.NavigateTo(Screen.Video);            break;
-                case 3: Menu.NavigateTo(Screen.Sound);            break;
-                case 4: Menu.NavigateTo(Screen.Main);             break;
-            }
-        }
-    }
-
-    private sealed class KeyboardBindingsHandler : ScreenHandler
-    {
-        public KeyboardBindingsHandler(MainMenuScreen menu) : base(menu) { }
-        public override string Title => Menu.RebindingAction != null
-            ? string.Format(Menu._localization.PressKeyTitle,
-                Menu._bindingActions.First(b => b.ConfigKey == Menu.RebindingAction).Label.ToUpper())
-            : Menu._localization.SettingsKeyboard.ToUpper();
-        public override int      ItemCount => Menu._bindingActions.Length;
-        public override string[] GetItems()
-            => Menu._bindingActions
-                .Select(b => b.ConfigKey == ""
-                    ? Menu._localization.Back
-                    : $"{b.Label,-8}  {Menu.KeyboardLabel(b.ConfigKey)}")
-                .ToArray();
-        public override void Activate(int index)
-        {
-            var (_, configKey) = Menu._bindingActions[index];
-            if (configKey == "")
-                Menu.NavigateTo(Screen.Settings);
-            else
-                Menu.RebindingAction = configKey;
-        }
-    }
-
-    private sealed class GamepadBindingsHandler : ScreenHandler
-    {
-        public GamepadBindingsHandler(MainMenuScreen menu) : base(menu) { }
-        public override string Title => Menu.GamepadRebindingAction != null
-            ? string.Format(Menu._localization.PressButtonTitle,
-                Menu._gamepadBindingActions.First(b => b.ConfigKey == Menu.GamepadRebindingAction).Label.ToUpper())
-            : Menu._localization.SettingsGamepad.ToUpper();
-        public override int      ItemCount => Menu._gamepadBindingActions.Length;
-        public override string[] GetItems()
-            => Menu._gamepadBindingActions
-                .Select(b => b.ConfigKey == ""
-                    ? Menu._localization.Back
-                    : $"{b.Label,-8}  {Menu.GetGamepadLabel(b.ConfigKey)}")
-                .ToArray();
-        public override bool IsItemEnabled(int index)
-        {
-            if (!SteamInputManager.IsUsingNativeActions()) return true;
-            var configKey = Menu._gamepadBindingActions[index].ConfigKey;
-            return configKey == "" || configKey == "OpenMenu";
-        }
-        public override void Activate(int index)
-        {
-            var (_, configKey) = Menu._gamepadBindingActions[index];
-            if (configKey == "")
-                Menu.NavigateTo(Screen.Settings);
-            else
-                Menu.GamepadRebindingAction = configKey;
-        }
-    }
-
-    private sealed class VideoHandler : ScreenHandler
-    {
-        public VideoHandler(MainMenuScreen menu) : base(menu) { }
-        public override string   Title     => Menu._localization.VideoTitle;
-        public override int      ItemCount => 4;
-        public override string[] GetItems() => new[]
-        {
-            Menu._config.WindowMode == "Fullscreen" ? Menu._localization.VideoWindowFullscreen : Menu._localization.VideoWindowWindowed,
-            Menu._config.GraphicsSmoothingEnabled   ? Menu._localization.VideoGraphicsSmooth   : Menu._localization.VideoGraphicsOriginal,
-            Menu._config.ShowFps                    ? Menu._localization.VideoFpsOn            : Menu._localization.VideoFpsOff,
-            Menu._localization.Back,
-        };
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case 0:
-                    Menu._onWindowModeToggle(Menu._config.WindowMode != "Fullscreen");
-                    break;
-                case 1:
-                    bool smoothOn = !Menu._config.GraphicsSmoothingEnabled;
-                    Menu._config.GraphicsSmoothingEnabled = smoothOn;
-                    Menu._onGraphicsScalerToggled(smoothOn);
-                    break;
-                case 2:
-                    Menu._config.ShowFps = !Menu._config.ShowFps;
-                    Menu._onConfigSaved();
-                    break;
-                case 3:
-                    Menu.NavigateTo(Screen.Settings);
-                    break;
-            }
-        }
-    }
-
-    private sealed class SoundHandler : ScreenHandler
-    {
-        public  const int VolumeIndex    = 0;
-        private const int ScrubberIndex  = 1;
-        private const int MusicIndex     = 2;
-        private const int BackIndex      = 3;
-        public SoundHandler(MainMenuScreen menu) : base(menu) { }
-        public override string   Title     => Menu._localization.SoundTitle;
-        public override int      ItemCount => 4;
-        public override string[] GetItems() => new[]
-        {
-            string.Format(Menu._localization.SoundVolume, Menu._config.Volume),
-            Menu._config.SoundScrubberEnabled  ? Menu._localization.SoundScrubberOn : Menu._localization.SoundScrubberOff,
-            Menu._config.MainMenuMusicEnabled  ? Menu._localization.SoundMusicOn    : Menu._localization.SoundMusicOff,
-            Menu._localization.Back,
-        };
-        public override void Activate(int index)
-        {
-            switch (index)
-            {
-                case ScrubberIndex:
-                    bool scrubOn = !Menu._config.SoundScrubberEnabled;
-                    Menu._config.SoundScrubberEnabled = scrubOn;
-                    Menu._onScrubberToggled(scrubOn);
-                    break;
-                case MusicIndex:
-                    bool musicOn = !Menu._config.MainMenuMusicEnabled;
-                    Menu._config.MainMenuMusicEnabled = musicOn;
-                    Menu._onMenuMusicToggled(musicOn);
-                    break;
-                case BackIndex:
-                    Menu.NavigateTo(Screen.Settings);
-                    break;
-            }
-        }
     }
 }
